refactor(ivy): move and rename `Analyzer` to `DecorationAnalyzer`
This is in preparation for adding in other kinds of Analyzer.
